Update test size and timestamp in forecasts.

+++ app/outers/interfaces/deliveries/contracts/requests/forecasts/item_stocks/stock_forecast_body.py
@@ -4,4 +4,4 @@ from pydantic import BaseModel
 class StockForecastBody(BaseModel):
     horizon: int
     resample: str
-    test_size: int
+    test_size: float
