+++ app/outers/interfaces/deliveries/contracts/requests/forecasts/item_stocks/stock_forecast_by_item_id_request.py
@@ -7,4 +7,4 @@ class StockForecastByItemIdRequest(BaseModel):
     item_id: UUID
     horizon: int
     resample: str
-    test_size: int
+    test_size: float
