+++ app/outers/interfaces/deliveries/contracts/requests/forecasts/item_transactions/transaction_forecast_body.py
@@ -4,4 +4,4 @@ from pydantic import BaseModel
 class TransactionForecastBody(BaseModel):
     horizon: int
     resample: str
-    test_size: int
+    test_size: float
