+++ app/outers/interfaces/deliveries/contracts/requests/forecasts/item_transactions/transaction_forecast_by_item_id_request.py
@@ -7,4 +7,4 @@ class TransactionForecastByItemIdRequest(BaseModel):
     item_id: UUID
     horizon: int
     resample: str
-    test_size: int
+    test_size: float
